requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pika==1.2.0
python-dotenv